Print packages

print_packages() is a Print function that prints output to console Specifically, this function implements an algorithm to print packages. The function is called for its side effects and does not return a value.

Usage

print_packages(
  pkg_extensions_tb = NULL,
  gh_repo_1L_chr = "ready4-dev/ready4",
  gh_tag_1L_chr = "Documentation_0.0",
  include_1L_chr = "modules",
  module_pkgs_chr = character(0),
  ns_var_nm_1L_chr = "pt_ns_chr",
  project_badges_url_1L_chr = "https://img.shields.io/badge/ready4",
  reference_var_nm_1L_chr = "Reference",
  scroll_height_1L_chr = character(0),
  scroll_width_1L_chr = character(0),
  url_stub_1L_chr = "https://ready4-dev.github.io/",
  vignette_var_nm_1L_chr = "Vignettes",
  vignette_url_var_nm_1L_chr = "Vignettes_URLs",
  what_chr = "all",
  ...
)

Arguments

pkg_extensions_tb

Package extensions (a tibble), Default: NULL

gh_repo_1L_chr

Github repository (a character vector of length one), Default: 'ready4-dev/ready4'

gh_tag_1L_chr

Github tag (a character vector of length one), Default: 'Documentation_0.0'

include_1L_chr

Include (a character vector of length one), Default: 'modules'

module_pkgs_chr

Module packages (a character vector), Default: character(0)

ns_var_nm_1L_chr

Namespace variable name (a character vector of length one), Default: 'pt_ns_chr'

project_badges_url_1L_chr

Project badges url (a character vector of length one), Default: 'https://img.shields.io/badge/ready4'

reference_var_nm_1L_chr

Reference variable name (a character vector of length one), Default: 'Reference'

scroll_height_1L_chr

Scroll height (a character vector of length one), Default: character(0)

scroll_width_1L_chr

Scroll width (a character vector of length one), Default: character(0)

url_stub_1L_chr

Url stub (a character vector of length one), Default: 'https://ready4-dev.github.io/'

vignette_var_nm_1L_chr

Vignette variable name (a character vector of length one), Default: 'Vignettes'

vignette_url_var_nm_1L_chr

Vignette url variable name (a character vector of length one), Default: 'Vignettes_URLs'

what_chr

What (a character vector), Default: 'all'

...

Additional arguments

Value

Package extensions (a kable)

Examples

print_packages(gh_repo_1L_chr = "ready4-dev/ready4",
               include_1L_chr = "framework")
#> <table class="table table-hover table-condensed" style="margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Type </th>
#>    <th style="text-align:left;"> Package </th>
#>    <th style="text-align:left;"> Purpose </th>
#>    <th style="text-align:left;"> Documentation </th>
#>    <th style="text-align:left;"> Code </th>
#>    <th style="text-align:left;"> Examples </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-authoring-maroon?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/ready4use/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Author, Ingest, Label and Share Health Economic Model Datasets </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4use/authors.html" style="     ">Citation</a>                                                           , <a href="https://ready4-dev.github.io/ready4use/index.html" style="     ">Website</a>                                                              , <a href="https://github.com/ready4-dev/ready4use/releases/download/Documentation_0.0/ready4use_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/ready4use/releases/download/Documentation_0.0/ready4use_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/ready4use/" style="     ">Dev</a>  , <a href="https://doi.org/10.5281/zenodo.5644336" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4use/articles/V_01.html" style="     ">1</a>, <a href="https://ready4-dev.github.io/ready4use/articles/V_02.html" style="     ">2</a>, <a href="https://ready4-dev.github.io/ready4use/articles/V_03.html" style="     ">3</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-authoring-maroon?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/ready4show/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Author Health Economic Analysis Programs and Reporting Templates </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4show/authors.html" style="     ">Citation</a>                                                            , <a href="https://ready4-dev.github.io/ready4show/index.html" style="     ">Website</a>                                                               , <a href="https://github.com/ready4-dev/ready4show/releases/download/Documentation_0.0/ready4show_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/ready4show/releases/download/Documentation_0.0/ready4show_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/ready4show/" style="     ">Dev</a> , <a href="https://doi.org/10.5281/zenodo.5644568" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4show/articles/V_01.html" style="     ">4</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-authoring-maroon?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/ready4fun/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Author and Document Functions that Implement Transferable Health </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4fun/authors.html" style="     ">Citation</a>                                                           , <a href="https://ready4-dev.github.io/ready4fun/index.html" style="     ">Website</a>                                                              , <a href="https://github.com/ready4-dev/ready4fun/releases/download/Documentation_0.0/ready4fun_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/ready4fun/releases/download/Documentation_0.0/ready4fun_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/ready4fun/" style="     ">Dev</a>  , <a href="https://doi.org/10.5281/zenodo.5611779" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4fun/articles/V_01.html" style="     ">5</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-authoring-maroon?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/ready4class/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Author and Document Classes that Make Health Economic Models </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4class/authors.html" style="     ">Citation</a>                                                             , <a href="https://ready4-dev.github.io/ready4class/index.html" style="     ">Website</a>                                                                , <a href="https://github.com/ready4-dev/ready4class/releases/download/Documentation_0.0/ready4class_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/ready4class/releases/download/Documentation_0.0/ready4class_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/ready4class/" style="     ">Dev</a>, <a href="https://doi.org/10.5281/zenodo.5640313" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4class/articles/V_01.html" style="     ">6</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-authoring-maroon?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/ready4pack/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Author R Packages of Health Economic Model Modules </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4pack/authors.html" style="     ">Citation</a>                                                            , <a href="https://ready4-dev.github.io/ready4pack/index.html" style="     ">Website</a>                                                               , <a href="https://github.com/ready4-dev/ready4pack/releases/download/Documentation_0.0/ready4pack_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/ready4pack/releases/download/Documentation_0.0/ready4pack_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/ready4pack/" style="     ">Dev</a> , <a href="https://doi.org/10.5281/zenodo.5644322" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4pack/articles/V_01.html" style="     ">7</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-foundation-coral?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/ready4/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Implement Transparent, Reusable and Updatable Computational Health </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4/authors.html" style="     ">Citation</a>                                                        , <a href="https://ready4-dev.github.io/ready4/index.html" style="     ">Website</a>                                                           , <a href="https://github.com/ready4-dev/ready4/releases/download/Documentation_0.0/ready4_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/ready4/releases/download/Documentation_0.0/ready4_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/ready4/" style="     ">Dev</a>     , <a href="https://doi.org/10.5281/zenodo.5606250" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/ready4/articles/V_01.html" style="     ">8</a> , <a href="https://ready4-dev.github.io/ready4/articles/V_02.html" style="     ">9</a> , <a href="https://ready4-dev.github.io/ready4/articles/V_03.html" style="     ">10</a>, <a href="https://ready4-dev.github.io/ready4/articles/V_04.html" style="     ">11</a> </td>
#>   </tr>
#> </tbody>
#> </table>
print_packages(gh_repo_1L_chr = "ready4-dev/ready4",
               include_1L_chr = "modules")
#> <table class="table table-hover table-condensed" style="margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Type </th>
#>    <th style="text-align:left;"> Package </th>
#>    <th style="text-align:left;"> Purpose </th>
#>    <th style="text-align:left;"> Documentation </th>
#>    <th style="text-align:left;"> Code </th>
#>    <th style="text-align:left;"> Examples </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-description-navy?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/youthvars/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Describe and Validate Youth Mental Health Dataset Variables </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/youthvars/authors.html" style="     ">Citation</a>                                                           , <a href="https://ready4-dev.github.io/youthvars/index.html" style="     ">Website</a>                                                              , <a href="https://github.com/ready4-dev/youthvars/releases/download/Documentation_0.0/youthvars_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/youthvars/releases/download/Documentation_0.0/youthvars_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/youthvars/" style="     ">Dev</a>  , <a href="https://doi.org/10.5281/zenodo.5646550" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/youthvars/articles/V_01.html" style="     ">12</a>, <a href="https://ready4-dev.github.io/youthvars/articles/V_02.html" style="     ">13</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-description-navy?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/scorz/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Score Multi-Attribute Utility Instruments </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/scorz/authors.html" style="     ">Citation</a>                                                       , <a href="https://ready4-dev.github.io/scorz/index.html" style="     ">Website</a>                                                          , <a href="https://github.com/ready4-dev/scorz/releases/download/Documentation_0.0/scorz_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/scorz/releases/download/Documentation_0.0/scorz_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/scorz/" style="     ">Dev</a>      , <a href="https://doi.org/10.5281/zenodo.5722708" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/scorz/articles/V_01.html" style="     ">14</a>, <a href="https://ready4-dev.github.io/scorz/articles/V_02.html" style="     ">15</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-modelling-indigo?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/mychoice/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Model Youth Choice Behaviours </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/mychoice/authors.html" style="     ">Citation</a>                                               , <a href="https://ready4-dev.github.io/mychoice/index.html" style="     ">Website</a>                                                  , <a href="https://github.com/ready4-dev/mychoice/releases/download/Documentation_0.0/mychoice_Developer.pdf" style="     ">Citation</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/mychoice/" style="     ">Dev</a>   , <a href="https://doi.org/10.5281/zenodo.7213799" style="     ">Archive</a> </td>
#>    <td style="text-align:left;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-modelling-indigo?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/TTU/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Implement Transfer to Utility Mapping Algorithms </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/TTU/authors.html" style="     ">Citation</a>                                                     , <a href="https://ready4-dev.github.io/TTU/index.html" style="     ">Website</a>                                                        , <a href="https://github.com/ready4-dev/TTU/releases/download/Documentation_0.0/TTU_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/TTU/releases/download/Documentation_0.0/TTU_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/TTU/" style="     ">Dev</a>        , <a href="https://doi.org/10.5281/zenodo.5646593" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/TTU/articles/V_01.html" style="     ">16</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-modelling-indigo?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/heterodox/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Explore and Characterise Heterogeneity in Quality of Life Data </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/heterodox/authors.html" style="     ">Citation</a>                                                           , <a href="https://ready4-dev.github.io/heterodox/index.html" style="     ">Website</a>                                                              , <a href="https://github.com/ready4-dev/heterodox/releases/download/Documentation_0.0/heterodox_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/heterodox/releases/download/Documentation_0.0/heterodox_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/heterodox/" style="     ">Dev</a>  , <a href="https://doi.org/10.5281/zenodo.5751193" style="     ">Archive</a> </td>
#>    <td style="text-align:left;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-modelling-indigo?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/specific/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Specify Models to Solve Inverse Problems </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/specific/authors.html" style="     ">Citation</a>                                                          , <a href="https://ready4-dev.github.io/specific/index.html" style="     ">Website</a>                                                             , <a href="https://github.com/ready4-dev/specific/releases/download/Documentation_0.0/specific_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/specific/releases/download/Documentation_0.0/specific_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/specific/" style="     ">Dev</a>   , <a href="https://doi.org/10.5281/zenodo.5768689" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/specific/articles/V_01.html" style="     ">17</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-prediction-forestgreen?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/youthu/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Transform Youth Outcomes to Health Utility Predictions </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/youthu/authors.html" style="     ">Citation</a>                                                        , <a href="https://ready4-dev.github.io/youthu/index.html" style="     ">Website</a>                                                           , <a href="https://github.com/ready4-dev/youthu/releases/download/Documentation_0.0/youthu_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/youthu/releases/download/Documentation_0.0/youthu_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/youthu/" style="     ">Dev</a>     , <a href="https://doi.org/10.5281/zenodo.5646668" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/youthu/articles/V_01.html" style="     ">18</a> </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-modelling-indigo?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/aus/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Model Australian Spatial Data </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/aus/authors.html" style="     ">Citation</a>                                                     , <a href="https://ready4-dev.github.io/aus/index.html" style="     ">Website</a>                                                        , <a href="https://github.com/ready4-dev/aus/releases/download/Documentation_0.0/aus_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/aus/releases/download/Documentation_0.0/aus_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/aus/" style="     ">Dev</a>        , <a href="https://doi.org/10.5281/zenodo.7687126" style="     ">Archive</a> </td>
#>    <td style="text-align:left;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-modelling-indigo?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/vicinity/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Model Spatial Features of Health Systems </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/vicinity/authors.html" style="     ">Citation</a>                                                          , <a href="https://ready4-dev.github.io/vicinity/index.html" style="     ">Website</a>                                                             , <a href="https://github.com/ready4-dev/vicinity/releases/download/Documentation_0.0/vicinity_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/vicinity/releases/download/Documentation_0.0/vicinity_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/vicinity/" style="     ">Dev</a>   , <a href="https://doi.org/10.5281/zenodo.7623630" style="     ">Archive</a> </td>
#>    <td style="text-align:left;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-modelling-indigo?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/bimp/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Undertake Health Economic Budget Impact Analysis. </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/bimp/authors.html" style="     ">Citation</a>                                                      , <a href="https://ready4-dev.github.io/bimp/index.html" style="     ">Website</a>                                                         , <a href="https://github.com/ready4-dev/bimp/releases/download/Documentation_0.0/bimp_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/bimp/releases/download/Documentation_0.0/bimp_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/bimp/" style="     ">Dev</a>       , <a href="https://doi.org/10.5281/zenodo.5889462" style="     ">Archive</a> </td>
#>    <td style="text-align:left;">  </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  <img src="https://img.shields.io/badge/ready4-modelling-indigo?style=flat&amp;labelColor=black&amp;logo=data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABHNCSVQICAgIfAhkiAAAAAFzUkdCAK7OHOkAAAAEZ0FNQQAAsY8L/GEFAAAACXBIWXMAABYlAAAWJQFJUiTwAAAAIXRFWHRDcmVhdGlvbiBUaW1lADIwMjI6MDM6MDcgMTY6MTM6NTPZeG5UAAABa0lEQVQ4T4WT607CQBCFpyUi3qIR0eAfNfCi/vENfEgENIAIlcJ6vr1oLaZOerJzdst0zpklc49nznqHZs6ZfWwDem1xM1sqXwtXkb8rL4SuOLEoLXPPXWfD01Dg9dPsrTQbngQ+EZ+LDyIfiy/FHyIfFZbbTslWKOOqxx/uWBPSfp07FahGlqlNfWGqL9HNfBO+CAfwdO55WS8g4MFML834sfJVA9e7vwsg50aGohncdmRojV9XeL+jArRNmZxVSJ4Acj3NHqARdyeFJqC2KJiCfKE9zsfxnNYTl5TcCtmNMcwY/ZXf+3wdzzVza2vj4iCaq3d1R/bvwVSH6IPjNIUHx0FSNZA7WquDqOVb35+eiO8h7Oe+vRfp0a3yGtFMDuiAIg2R20YaVwJ3Hj+4kehO/J/I7VJ/jHtpvBP6mrHnR4EzdyQ0xI8HhM8jUiChxVpDK3iVuadzx43yRdI4E2d0gNtX74TCs419AR8YEST/cHPBAAAAAElFTkSuQmCC">
#> </td>
#>    <td style="text-align:left;">  <img src="https://ready4-dev.github.io/costly/logo.png" width="51.2" height="51.2">
#> </td>
#>    <td style="text-align:left;"> Develop, Use and Share Unit Cost Datasets for Health Economic </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/costly/authors.html" style="     ">Citation</a>                                                        , <a href="https://ready4-dev.github.io/costly/index.html" style="     ">Website</a>                                                           , <a href="https://github.com/ready4-dev/costly/releases/download/Documentation_0.0/costly_User.pdf" style="     ">Manual - Short (PDF)</a>    , <a href="https://github.com/ready4-dev/costly/releases/download/Documentation_0.0/costly_Developer.pdf" style="     ">Manual - Full (PDF)</a> </td>
#>    <td style="text-align:left;"> <a href="https://github.com/ready4-dev/costly/" style="     ">Dev</a>    , <a href="https://github.com/ready4-dev/costly/" style="     ">Archive</a> </td>
#>    <td style="text-align:left;"> <a href="https://ready4-dev.github.io/costly/articles/V_01.html" style="     ">19</a>, <a href="https://ready4-dev.github.io/costly/articles/V_02.html" style="     ">20</a> </td>
#>   </tr>
#> </tbody>
#> </table>