Rowbind tibbles in ready4 S4 object

rowbind_tbs_in_r4_obj() is a Rowbind function that performs custom rowbind operations on table objects. Specifically, this function implements an algorithm to rowbind tibbles in ready4 s4 object. The function returns Tibbles (a ready4 S4).

rowbind_tbs_in_r4_obj(tbs_r4, slot_nm_1L_chr, second_tbs_r4, r4_name_1L_chr)

Arguments

tbs_r4

Tibbles (a ready4 S4)

slot_nm_1L_chr

Slot name (a character vector of length one)

second_tbs_r4

Second tibbles (a ready4 S4)

r4_name_1L_chr

Ready4 S4 name (a character vector of length one)

Value

Tibbles (a ready4 S4)